chore(gpu): refactor oprf test, remove unused arg and fix multi-GPU for oprf
