+++ backends/tfhe-cuda-backend/src/bindings.rs
@@ -1551,7 +1551,6 @@ unsafe extern "C" {
         pbs_base_log: u32,
         grouping_factor: u32,
         num_blocks_to_process: u32,
-        num_blocks: u32,
         message_modulus: u32,
         carry_modulus: u32,
         pbs_type: PBS_TYPE,
