+++ tfhe/src/integer/gpu/mod.rs
@@ -2785,7 +2785,6 @@ pub(crate) unsafe fn grouped_oprf_async<B: Numeric>(
     radix_lwe_out: &mut CudaRadixCiphertext,
     seeded_lwe_input: &CudaVec<u64>,
     num_blocks_to_process: u32,
-    total_num_blocks: u32,
     bootstrapping_key: &CudaVec<B>,
     lwe_dimension: LweDimension,
     glwe_dimension: GlweDimension,
@@ -2845,7 +2844,6 @@ pub(crate) unsafe fn grouped_oprf_async<B: Numeric>(
         pbs_base_log.0 as u32,
         grouping_factor.0 as u32,
         num_blocks_to_process,
-        total_num_blocks,
         message_modulus.0 as u32,
         carry_modulus.0 as u32,
         pbs_type as u32,
@@ -2874,7 +2872,6 @@ pub(crate) unsafe fn grouped_oprf_async<B: Numeric>(
 pub(crate) fn get_grouped_oprf_size_on_gpu(
     streams: &CudaStreams,
     num_blocks_to_process: u32,
-    total_num_blocks: u32,
     lwe_dimension: LweDimension,
     glwe_dimension: GlweDimension,
     polynomial_size: PolynomialSize,
@@ -2907,7 +2904,6 @@ pub(crate) fn get_grouped_oprf_size_on_gpu(
             pbs_base_log.0 as u32,
             grouping_factor.0 as u32,
             num_blocks_to_process,
-            total_num_blocks,
             message_modulus.0 as u32,
             carry_modulus.0 as u32,
             pbs_type as u32,
