+++ tfhe/src/integer/gpu/server_key/radix/tests_unsigned/mod.rs
@@ -6,6 +6,7 @@ pub(crate) mod test_div_mod;
 pub(crate) mod test_ilog2;
 pub(crate) mod test_mul;
 pub(crate) mod test_neg;
+pub(crate) mod test_oprf;
 pub(crate) mod test_rotate;
 pub(crate) mod test_scalar_add;
 pub(crate) mod test_scalar_bitwise_op;
